refactor how we write integration tests
