+++ test/integration/app/knexapp/package.json
